[all] Update to `swf-tree@0.5.0`
This commit updates the parser to use the new unified representation for edges. The presence of a control point determines the type of the edge. In the case of curved edges, `delta` is the difference between the start and end of the edges, it corresponds to the sum of the old `control_delta` and `anchor_delta`.