[all] Refactor tests
Refactor the tests to share them between both the Typescript and Rust implementation. Apart from the existing `standalone-movies` tests, there are now two families of tests: `tags` and `various`. Each of them contains groups of samples and each group a directory for each test. Each test has two files: `input.bytes` and `value.json`. Some of the inputs are authored directly while others are expressed in a textual format.

The textual format allows to represent hex or bin data. `#` indicates a comment until the end of the line. To input binary, the file must start with `# bin\n`. Invalid characters are ignored, allowing for punctuation and newlines in the file. Use the `build.js` files to convert the text files to bytes.