[rs] Add initial streaming parser
This commit adds an initial version of an SWF streaming parser. The test gives an example how to use it. This parser uses the _type state_ pattern to ensure the user first consumes the SWF header before parsing the tags.
There are some redundancies in the code and the error handling should be improved in further commits.
The streaming parser does not support LZMA yet. `lzma-rs` currently requires the inpput to be fully buffered.